+++ src/main/java/hexlet/code/service/TaskStatusService.java
@@ -29,7 +29,7 @@ public class TaskStatusService {
 
     public TaskStatusDTO getTaskStatusById(Long id) {
         return taskStatusMapper.map(taskStatusRepository.findById(id)
-                .orElseThrow(() -> new ResourceNotFoundException("TaskStatus not found")));
+                .orElseThrow(() -> new ResourceNotFoundException("TaskStatus with id " + id + " not found!")));
     }
 
     public TaskStatusDTO createTaskStatus(TaskStatusCreateDTO taskStatusDTO) {
@@ -42,14 +42,18 @@ public class TaskStatusService {
 
     public TaskStatusDTO updateTaskStatus(long id, TaskStatusUpdateDTO taskStatusDTO) {
         var taskStatus = taskStatusRepository.findById(id)
-                .orElseThrow(() -> new ResourceNotFoundException("TaskStatus not found"));
+                .orElseThrow(() -> new ResourceNotFoundException("TaskStatus with id " + id + " not found!"));
         taskStatusMapper.update(taskStatusDTO, taskStatus);
         return taskStatusMapper.map(taskStatusRepository.save(taskStatus));
     }
 
     public void deleteTaskStatus(long id) {
         var taskStatus = taskStatusRepository.findById(id)
-                .orElseThrow(() -> new ResourceNotFoundException("TaskStatus not found"));
+                .orElseThrow(() -> new ResourceNotFoundException("TaskStatus with id " + id + " not found!"));
+        // Если статус связан хотя бы с одной задачей, его нельзя удалить.
+        if (!taskStatus.getTasks().isEmpty()) {
+            throw new RuntimeException("Can't delete status, because it has tasks");
+        }
         taskStatusRepository.delete(taskStatus);
     }
 }
